requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
pycparser==2.21
python-telegram-bot==21.11
PyYAML==6.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.8
pytest~=7.3
pylint~=2.17
